requirements.txt
#PyQt5==5.15
PyQt6==6.7
python_dateutil==2.9
webuntis==0.1
requests~=2.32.3
python-dateutil~=2.9.0